LayoutTests/fast/dom/navigator-vibration.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test for the Vibration API.');

shouldBe("navigator.webkitVibrate(0);", "undefined");
shouldBe("navigator.webkitVibrate([]);", "undefined");
shouldBe("navigator.webkitVibrate(1000);", "undefined");
shouldBe("navigator.webkitVibrate([1000, 300, 500]);", "undefined");
shouldBe("navigator.webkitVibrate(4294967295);", "undefined");
shouldThrow("navigator.webkitVibrate();");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
